+++ include/normalize.h
@@ -2,5 +2,5 @@
 
 #ifndef NORMALIZE_H_INCLUDED
 #define NORMALIZE_H_INCLUDED
-void normalize(model_t *model, float** Test_Vect, int pos);
+void normalize(model_t *model, double** Test_Vect, int pos);
 #endif
