+++ include/test_matrix.h
@@ -1,8 +1,8 @@
-#include "model.h"
-
 #ifndef TESTMATRIX_H_INCLUDED
 #define TESTMATRIX_H_INCLUDED
 
-int test_matrix(model_t* foo_mat, double **Test_Vect, int T, int NClasses, int *Align, float *Pvit_w);
+#include "model.h"
+
+int test_matrix(model_t* foo_mat, double **Test_Vect, int T, int NClasses, int *Align, double *Pvit_w);
 
 #endif
