+++ include/viterbi_cycle.h
@@ -1,6 +1,6 @@
 #ifndef VITERBICYCLE_H_INCLUDED
 #define VITERBICYCLE_H_INCLUDED
 
-void viterbi_cycle(float *FI, int *XX, float *A_Vect, float *LogPdf, int N,  int T);
+void viterbi_cycle(double *FI, int *XX, double *A_Vect, double *LogPdf, int N, int T);
 
 #endif
