+++ include/viterbi_end.h
@@ -1,8 +1,8 @@
 #ifndef VITERBIEND_H_INCLUDED
 #define VITERBIEND_H_INCLUDED
 
-void maxx(float *Ar, int dim, float *val, int *pos);
+void maxx(double *Ar, int dim, double *val, int *pos);
 
-void viterbi_end(float *FI, int *XX, float *A_Vect, float *LogPdf, float *Pvit, int N, int T);
+void viterbi_end(double *FI, int *XX, double *A_Vect, double *LogPdf, double *Pvit, int N, int T);
 
 #endif
